Fix duplicate exclusion argument entries
It happens that when backup paths contains the homedir of a user,
exclusion file is computed by duplicate: one time by being a backup path
and another time for being a homedir.multi-repo

Having duplicate arguments seems to make no harm (other than probably
making restic parser do the job twice), but is weird and probably not
desirable.

This solution is not ideal but a workaround for now.