Add upgrade target to the Makefile
`make upgrade` will help to upgrade the installation with new features
and/or breaking configuration changes. The upgrade process will keep a
copy (backup) of config files and make a best effort to restore previous
values.

User manual action should be not required for default settings, and be
minimal in the case of customizations.